+++ include/ecal/viz/event_render.hpp
@@ -9,7 +9,7 @@
 namespace ecal::viz {
 
 cv::Mat eventsToImage(const std::vector<ecal::core::TimedEventNs> &events,
-                      int width, int height, int zoom_factor = 1,
+                      int width, int height, float zoom_factor = 1.0f,
                       bool non_max_suppression = false, int add_value = 50);
 
 } // namespace ecal::viz
